requirements.txt
aiogram==3.17.0
python-dotenv==1.0.0
openai==1.57.4
requests==2.31.0
tiktoken==0.8.0